# Conflicts:
#	src/main/java/sim/danslchamp/controleurs/CircuitControleur.java

+++ src/main/java/sim/danslchamp/circuit/Source.java
@@ -19,7 +19,7 @@ public abstract class Source extends Composant {
     }
 
     @Override
-    public double calculResistance(int frequence) {
+    public double calculResistance(double frequence) {
         return 0;
     }
 }
